GENERICATT-110: Add a telephone number entry type
Relies on com.googlecode.libphonenumber for telephone number managment.
The telephone number is parsed and displayed according to the region specified at
question creation time (for instance FR for french numbers, or IT for italian
numbers).
The telephone number is store normalized (E164, which is international format
with no formatting).